# 8265491: Math Signum optimization for x86