--- a/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
+++ b/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
@@ -1045,6 +1045,42 @@ void C2_MacroAssembler::evminmax_fp(int opcode, BasicType elem_bt,
   }
 }
 
+// Float/Double signum
+void C2_MacroAssembler::signum_fp(int opcode, XMMRegister dst,
+                                  XMMRegister zero, XMMRegister one,
+                                  Register scratch) {
+  assert(opcode == Op_SignumF || opcode == Op_SignumD, "sanity");
+
+  Label DONE_LABEL;
+
+  if (opcode == Op_SignumF){
+    assert(UseSSE > 0, "required");
+    ucomiss(dst, zero);
+  } else if (opcode == Op_SignumD){
+    assert(UseSSE > 1, "required");
+    ucomisd(dst, zero);
+  }
+
+  jcc(Assembler::equal, DONE_LABEL);
+  jcc(Assembler::parity, DONE_LABEL);
+
+  if (opcode == Op_SignumF){
+    movflt(dst, one);
+  } else if (opcode == Op_SignumD){
+    movdbl(dst, one);
+  }
+
+  jcc(Assembler::above, DONE_LABEL);
+
+  if (opcode == Op_SignumF){
+    xorps(dst, ExternalAddress(StubRoutines::x86::vector_float_sign_flip()), scratch);
+  } else if (opcode == Op_SignumD){
+    xorpd(dst, ExternalAddress(StubRoutines::x86::vector_double_sign_flip()), scratch);
+  }
+
+  bind(DONE_LABEL);
+}
+
 void C2_MacroAssembler::vextendbw(bool sign, XMMRegister dst, XMMRegister src) {
   if (sign) {
     pmovsxbw(dst, src);