--- a/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
+++ b/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
@@ -1061,8 +1061,8 @@ void C2_MacroAssembler::signum_fp(int opcode, XMMRegister dst,
     ucomisd(dst, zero);
   }
 
-  jcc(Assembler::equal, DONE_LABEL);
-  jcc(Assembler::parity, DONE_LABEL);
+  jcc(Assembler::equal, DONE_LABEL);    // handle special case +0.0/-0.0, if argument is +0.0/-0.0, return argument
+  jcc(Assembler::parity, DONE_LABEL);   // handle special case NaN, if argument NaN, return NaN
 
   if (opcode == Op_SignumF){
     movflt(dst, one);
@@ -1073,9 +1073,9 @@ void C2_MacroAssembler::signum_fp(int opcode, XMMRegister dst,
   jcc(Assembler::above, DONE_LABEL);
 
   if (opcode == Op_SignumF){
-    xorps(dst, ExternalAddress(StubRoutines::x86::vector_float_sign_flip()), scratch);
+    xorps(dst, ExternalAddress(StubRoutines::x86::float_sign_flip()), scratch);
   } else if (opcode == Op_SignumD){
-    xorpd(dst, ExternalAddress(StubRoutines::x86::vector_double_sign_flip()), scratch);
+    xorpd(dst, ExternalAddress(StubRoutines::x86::double_sign_flip()), scratch);
   }
 
   bind(DONE_LABEL);