--- a/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
+++ b/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
@@ -1053,28 +1053,21 @@ void C2_MacroAssembler::signum_fp(int opcode, XMMRegister dst,
 
   Label DONE_LABEL;
 
-  if (opcode == Op_SignumF){
+  if (opcode == Op_SignumF) {
     assert(UseSSE > 0, "required");
     ucomiss(dst, zero);
-  } else if (opcode == Op_SignumD){
+    jcc(Assembler::equal, DONE_LABEL);    // handle special case +0.0/-0.0, if argument is +0.0/-0.0, return argument
+    jcc(Assembler::parity, DONE_LABEL);   // handle special case NaN, if argument NaN, return NaN
+    movflt(dst, one);
+    jcc(Assembler::above, DONE_LABEL);
+    xorps(dst, ExternalAddress(StubRoutines::x86::vector_float_sign_flip()), scratch);
+  } else if (opcode == Op_SignumD) {
     assert(UseSSE > 1, "required");
     ucomisd(dst, zero);
-  }
-
-  jcc(Assembler::equal, DONE_LABEL);    // handle special case +0.0/-0.0, if argument is +0.0/-0.0, return argument
-  jcc(Assembler::parity, DONE_LABEL);   // handle special case NaN, if argument NaN, return NaN
-
-  if (opcode == Op_SignumF){
-    movflt(dst, one);
-  } else if (opcode == Op_SignumD){
+    jcc(Assembler::equal, DONE_LABEL);    // handle special case +0.0/-0.0, if argument is +0.0/-0.0, return argument
+    jcc(Assembler::parity, DONE_LABEL);   // handle special case NaN, if argument NaN, return NaN
     movdbl(dst, one);
-  }
-
-  jcc(Assembler::above, DONE_LABEL);
-
-  if (opcode == Op_SignumF){
-    xorps(dst, ExternalAddress(StubRoutines::x86::vector_float_sign_flip()), scratch);
-  } else if (opcode == Op_SignumD){
+    jcc(Assembler::above, DONE_LABEL);
     xorpd(dst, ExternalAddress(StubRoutines::x86::vector_double_sign_flip()), scratch);
   }