--- a/src/hotspot/cpu/x86/vm_version_x86.cpp
+++ b/src/hotspot/cpu/x86/vm_version_x86.cpp
@@ -1700,7 +1700,7 @@ void VM_Version::get_processor_features() {
     }
   }
 #endif // !PRODUCT
-  if (FLAG_IS_DEFAULT(UseSignumIntrinsic) && (UseSSE >= 2)) {
+  if (FLAG_IS_DEFAULT(UseSignumIntrinsic)) {
       FLAG_SET_DEFAULT(UseSignumIntrinsic, true);
   }
 }