--- a/src/hotspot/cpu/x86/x86.ad
+++ b/src/hotspot/cpu/x86/x86.ad
@@ -5778,7 +5778,7 @@ instruct evminmaxFP_reg_eavx(vec dst, vec a, vec b, vec atmp, vec btmp, kReg ktm
 // --------------------------------- Signum ---------------------------
 
 instruct signumF_reg(regF dst, regF zero, regF one, rFlagsReg cr) %{
-  predicate(UseSSE>=1)
+  predicate(UseSSE>=1);
   match(Set dst (SignumF dst (Binary zero one)));
   effect(KILL cr);
   format %{ "signumF $dst, $dst" %}
@@ -5798,7 +5798,7 @@ instruct signumF_reg(regF dst, regF zero, regF one, rFlagsReg cr) %{
 %}
 
 instruct signumD_reg(regD dst, regD zero, regD one, rFlagsReg cr) %{
-  predicate(UseSSE>=2)
+  predicate(UseSSE>=2);
   match(Set dst (SignumD dst (Binary zero one)));
   effect(KILL cr);
   format %{ "signumD $dst, $dst" %}