--- a/src/hotspot/cpu/x86/x86.ad
+++ b/src/hotspot/cpu/x86/x86.ad
@@ -5790,8 +5790,7 @@ instruct signumF_reg(regF dst, regF zero, regF one, rFlagsReg cr) %{
     __ jcc(Assembler::equal, exit);
     __ movflt($dst$$XMMRegister, $one$$XMMRegister);
     __ jcc(Assembler::above, exit);
-    __ movflt($dst$$XMMRegister, $zero$$XMMRegister);
-    __ subss($dst$$XMMRegister, $one$$XMMRegister);
+    __ xorps($dst$$XMMRegister, ExternalAddress(float_signflip()));
     __ bind(exit);
   %}
   ins_pipe( pipe_slow );
@@ -5810,8 +5809,7 @@ instruct signumD_reg(regD dst, regD zero, regD one, rFlagsReg cr) %{
     __ jcc(Assembler::equal, exit);
     __ movdbl($dst$$XMMRegister, $one$$XMMRegister);
     __ jcc(Assembler::above, exit);
-    __ movdbl($dst$$XMMRegister, $zero$$XMMRegister);
-    __ subsd($dst$$XMMRegister, $one$$XMMRegister);
+    __ xorpd($dst$$XMMRegister, ExternalAddress(double_signflip()));
     __ bind(exit);
   %}
   ins_pipe( pipe_slow );