@mgkwill 2021-04-22
- Switched to 3 operand xorps/d on colleague suggestion to insure I don't introduce a hard to find register bug. 

- Implemented @DamonFool suggestion for Default Enable of UseSignumIntrinsic on x86. Slightly worried about `!has_match_rule(opcode)` for aarch64. But as long as it doesn't hit this I think it will work.

 src/hotspot/cpu/aarch64/aarch64.ad:
```
const bool Matcher::match_rule_supported(int opcode) {
  if (!has_match_rule(opcode))
    return false;
```

Also moved matcher checks in src/hotspot/cpu/x86/x86.ad outside of 
```
#ifndef _LP64
#endif // !LP64
```

Thanks for the review @DamonFool @dholmes-ora @jatin-bhateja.