@mgkwill 2021-04-23
> Looks good to me.
> Thanks for your update.
> 
> Please make sure the copyright year has been updated (TestSignumIntrinsic.java?) and all the imports are actually needed (import org.openjdk.jmh.annotations.Setup; ?).

Thanks @DamonFool. Updated with your suggestions.