@mgkwill 2021-04-27
Thanks for the review @sviswa7. 

I've updated to `StubRoutines::x86::float_sign_flip()` & `StubRoutines::x86::double_sign_flip()` as I was using when the code was in the src/hotspot/cpu/x86/x86.ad. If I understood your comment correctly this should use 128 bit aligned.

I've added comments as suggested.